// DXC_Frontend/src/api/models/tinTucCategoryDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface TinTucCategoryDto {
  id?: number
  publicId?: string
  /** @nullable */
  name: string | null
  /** @nullable */
  slug: string | null
  /** @nullable */
  description?: string | null
  displayOrder?: number
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
}